Add placeholders for ew-ver and ew-rel categories
These are initial tests that serve as a starting point. More should be
added in the future.

The 'ew-ver' test added should work for what is described on the CESM
website for held-suarez. This is a 1200 day forward run of the FHS94
compset with the 120km 32 level MPAS-A grid.
See "https://www.cesm.ucar.edu/models/simple/held-suarez"

The 'ew-rel' tests are starting with FullyCoupled-EW runs on 120, 60,
30, and 15km grids. These tests are "exact restart" tests that run
forward for 11 days, write a restart on day 6, and then restarts with
half the number of tasks to reach the same 11 day mark.