Add a testmod_dir to enable some simple data check tests
Tests that will use the ew-data_check testmod should be run with
'--no-setup' to keep the impact as light as possible. Since failure of
shell_commands during the CREATE_NEWCASE step are ignored, the success
of './check_input_data' is recorded in the ew_check_data.txt file.

A supplemental script is recommended to further automate this test.