Add 9 step exact restart tests for FHS94, F2000climoEW, and CHAOS2000dev
These tests (ERP) write a restart file part way through, and then
restart with half the number of processor elements (PEs). The tests
added in this commit makes sure restarts are possible for the simplest
compset (FHS94) with all compilers, for a compset using the prescribed
mode from MPAS-SeaIce with nvhpc, and for the compset using
MPAS-{A,O,SI} with all compilers.

These tests will replace the SMS tests in a future commit.